refactor(services): consolidate email sender into @revealui/services/email
Migrates the parallel Gmail JWT/REST implementations from
apps/server/src/lib/email.ts (~210 LOC functional) and
apps/admin/src/lib/email/index.ts (~250 LOC class-based) into a single
shared package at packages/services/src/email/.

Design picks the admin's class-based shape — more general, extensible
to additional providers (SES, Resend, fallback chains) without re-doing
the abstraction work in each app:

- EmailProvider interface + GmailProvider class + MockEmailProvider
- sendEmail(options, { maxRetries, logger }) with exponential-backoff
  retry; throws in NODE_ENV=production after all retries fail
- Optional logger injection for apps that need a non-default
  observability sink (defaults to @revealui/utils/logger)
- sanitizeEmailHeader exported for reuse

App wrappers preserved:
- apps/server/src/lib/email.ts: thin wrapper preserving the historical
  void-returning + throw-in-prod API. Passes server's Hono observability
  logger through to the shared retry path. Existing email.test.ts
  continues to assert void return + throw-in-prod, both still pass.
- apps/admin/src/lib/email/index.ts: re-exports the shared primitives
  and keeps the admin-specific senders (sendRecoveryEmail,
  sendPasswordResetEmail) that bake in admin URL config.
- apps/admin/src/lib/email/verification.ts: unchanged (still imports
  sendEmail from ./index, which now resolves to the shared impl).

packages/services dependencies bumped: +jose ^6.2.2 (Gmail JWT signing),
+@revealui/utils workspace:* (default logger).

Closes the audit-violations/per-app-utils-overlap.md F1 finding.
